+++ src/main/java/act/inject/param/CliContextParamLoader.java
@@ -173,7 +173,15 @@ public class CliContextParamLoader extends ParamValueLoaderService {
             Class<? extends StringValueResolver>[] resolvers = resolve.value();
             for (Class<? extends StringValueResolver> resolverClass : resolvers) {
                 StringValueResolver resolver = injector.get(resolverClass);
-                if (rawType.isAssignableFrom(resolver.targetType())) {
+                Class<?> targetType = resolver.targetType();
+                boolean matches = rawType.isAssignableFrom(targetType);
+                if (!matches) {
+                    Class<?> rawType2 = $.wrapperClassOf(rawType);
+                    if (rawType != rawType2) {
+                        matches = rawType2.isAssignableFrom(targetType);
+                    }
+                }
+                if (matches) {
                     return resolver;
                 }
             }
@@ -191,7 +199,15 @@ public class CliContextParamLoader extends ParamValueLoaderService {
                 for (Class<? extends StringValueResolver> resolverClass : resolvers) {
                     StringValueResolver resolver = injector.get(resolverClass);
                     resolver.attributes($.evaluate(a));
-                    if (rawType.isAssignableFrom(resolver.targetType())) {
+                    Class<?> targetType = resolver.targetType();
+                    boolean matches = rawType.isAssignableFrom(targetType);
+                    if (!matches) {
+                        Class<?> rawType2 = $.wrapperClassOf(rawType);
+                        if (rawType != rawType2) {
+                            matches = rawType2.isAssignableFrom(targetType);
+                        }
+                    }
+                    if (matches) {
                         return resolver;
                     }
                 }