+++ src/main/java/act/util/DataObjectEnhancer.java
@@ -37,7 +37,6 @@ public class DataObjectEnhancer extends AppByteCodeEnhancer<DataObjectEnhancer> 
     private ObjectMetaInfo metaInfo;
 
     public DataObjectEnhancer() {
-        super(S.F.startsWith("act.").negate().or(S.F.startsWith("act.fsa")));
     }
 
     protected DataObjectEnhancer(ClassVisitor cv) {