+++ testapp/src/test/java/testapp/endpoint/GHIssue297.java
@@ -22,7 +22,7 @@ public class GHIssue297 extends EndpointTester {
 
     @Test(expected = NotFound.class)
     public void testFoo() throws Exception {
-        Map<String, Object> payload = C.map("person", new ObjectId(), "list", new String[] {"1", "a"});
+        Map<String, Object> payload = C.Map("person", new ObjectId(), "list", new String[] {"1", "a"});
         url("/gh/297/").postJSON(payload);
         checkRespCode();
     }