+++ testapp/src/test/java/testapp/endpoint/GHIssue317.java
@@ -14,7 +14,7 @@ public class GHIssue317 extends EndpointTester {
         String prodId = "Car 007";
         GH317.Product product = new GH317.Product(prodId);
         int quantity = 3;
-        Map<String, Object> order = C.map("prod", prodId, "quantity", 3);
+        Map<String, Object> order = C.Map("prod", prodId, "quantity", 3);
         url("/gh/317").postJSON(order).accept(H.Format.JSON);
         String s = resp().body().string();
         System.out.println(s);